not quite working, but close.  New logo too
